requirements.txt
# MCP相关依赖 - 从GitHub安装最新版本
# pip install git+https://github.com/modelcontextprotocol/python-sdk.git

# 数据处理和验证
pydantic>=2.0.0

# 富途API (用于实际连接)
futu-api>=6.0.0

# Web框架相关 (MCP依赖)
anyio>=4.5
httpx>=0.27
httpx-sse>=0.4
pydantic-settings>=2.5.2
python-multipart>=0.0.9
sse-starlette>=1.6.1
starlette>=0.27
uvicorn>=0.23.1 